requirements.txt
# Deep Learning
tensorflow==2.17.0

# Data Processing
numpy>=1.26.0,<2.0.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python-headless>=4.8.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Utilities
scikit-learn>=1.3.0